Include card's types (main unit, subunit, year)
Also added colors/formatted strings for all six subunits under the two
original groups. PDP is still not listed because it's not really in the
game as of yet. (At time of writing, there are just solo songs for each
of the nine PDP girls on both main groups' B-sides, nothing else.)