Add some more "nicknames", some of which aren't really
We're all tired of searching for "umi" and getting Hanayo or some random
N girl because her name contains the letters U, M, and I in exactly the
right order, in the middle of a word.

So, in addition to adding Eli's official in-universe nickname (romanized
both ways, just for completeness), we'll also try to make queries behave
as expected even when they're not quite specific enough.

Oh, and I guess the Number One Idol in the Universe also needs both of
her romanization options handled, too.