Nuke alternative kizuna calculation method
I thought the performance of _bond_loop() would be bad, but not THIS
bad. Turns out the difference is about 400x for a large combo value.

Initially I tried timing it with the default 1,000,000 iterations
(thought it was 10k, actually) and the loop-based method took so long
that I eventually Ctrl-C'd it.

>>> timeit.timeit("_bond_equation(901)", setup=setup, number=10000)
0.003494699951261282
>>> timeit.timeit("_bond_loop(901)", setup=setup, number=10000)
1.3935365000506863

So yeah. Elegant as the loop version is, compared to the equation, no
amount of code elegance is worth sacrificing THAT much performance.

Goodbye, _bond_loop(), we hardly knew ye.